+++ frontend/src/lib/ConnectionFavicon.svelte
@@ -36,10 +36,10 @@
 	import faviconDoujinMusicInfo from './connection_favicon/同人音楽info.png';
 	import faviconOtomadWiki2 from './connection_favicon/音MAD Wiki 2.png';
 	import faviconSoundcloud from './connection_favicon/Soundcloud.png';
-	import { TagWorkConnection } from '$lib/TagWorkConnection.js';
-	import type { MediaConnection } from '$lib/MediaConnection';
-	import type { SongConnection } from '$lib/SongConnection';
-	import type { ProfileConnection } from '$lib/ProfileConnection';
+	import type { TagWorkConnection } from '$lib/enums/TagWorkConnection';
+	import type { MediaConnection } from '$lib/enums/MediaConnection';
+	import type { SongConnection } from '$lib/enums/SongConnection';
+	import type { ProfileConnection } from '$lib/enums/ProfileConnection';
 
 	type FaviconKey =
 		| (typeof TagWorkConnection)[keyof typeof TagWorkConnection]['name']
