+++ frontend/src/lib/GuidelineWarning.svelte
@@ -4,7 +4,7 @@
 	import { m } from './paraglide/messages';
 	import { getLocale } from './paraglide/runtime';
 	import { FAQ_POST_ID, GUIDELINE_POST_ID } from './ui';
-	import { isSOV, isSVO } from './Languages';
+	import { isSOV, isSVO } from '$lib/enums/Languages';
 
 	let r = null;
