+++ frontend/src/lib/WorkTag.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { getTagDisplayName } from './api';
-	import { resolveCreatorRoleById } from './CreatorRole';
-	import { resolveWorkTagCategoryKeyById, WorkTagCategory } from './WorkTagCategory';
+	import { creatorRole, resolveCreatorRoleKeyById } from '$lib/enums/CreatorRole';
+	import { resolveWorkTagCategoryKeyById, WorkTagCategory } from '$lib/enums/WorkTagCategory';
 
 	interface Props {
 		tag: {
@@ -9,7 +9,7 @@
 			slug: string;
 			category: Parameters<typeof resolveWorkTagCategoryKeyById>[0];
 			sample?: boolean;
-			creator_roles?: Parameters<typeof resolveCreatorRoleById>[0][] | null;
+			creator_roles?: Parameters<typeof resolveCreatorRoleKeyById>[0][] | null;
 		};
 		selected?: boolean;
 		onClick?: (tag: Props['tag']) => void;
@@ -40,7 +40,7 @@
 </a>
 {#if category === 'CREATOR' && tag.creator_roles?.length}
 	<address class="text-otodb-content-fainter inline px-1 text-xs">
-		{#each tag.creator_roles as role, i (i)}{resolveCreatorRoleById(role).nameFn()}
+		{#each tag.creator_roles as role, i (i)}{creatorRole[resolveCreatorRoleKeyById(role)].nameFn()}
 			{#if i < tag.creator_roles.length - 1},&nbsp{/if}
 		{/each}
 	</address>
