+++ frontend/src/lib/enums/UserLevel.ts
