+++ frontend/src/lib/ui.ts
@@ -1,7 +1,7 @@
 import { browser } from '$app/environment';
 import client from './api';
 import { Languages, themes, type SupportedLanguage } from './enums';
-import { languages } from './Languages';
+import { languages } from './enums/Languages';
 import { getLocale, setLocale } from './paraglide/runtime';
 import { enhance } from '$app/forms';
 import { m } from './paraglide/messages';
