+++ frontend/src/routes/song_attribute/alias/+page.svelte
@@ -4,7 +4,7 @@
 	import TagsField from '$lib/TagsField.svelte';
 	import client from '$lib/api';
 	import { goto } from '$app/navigation';
-	import { isSOV, isSVO } from '$lib/Languages';
+	import { isSOV, isSVO } from '$lib/enums/Languages';
 	import { getLocale } from '$lib/paraglide/runtime';
 	import GuidelineWarning from '$lib/GuidelineWarning.svelte';
